Split persistent_index.hpp into helpers and persistent_index
